Added some updates for usage with Event Streams v10
Added a few updated yamls in an Event Streams v10 section. I'm not sure how you want to format it and leave it as how I had it or to maybe refactor it into their own respective addons to the previous sections.

Also, I'm unsure if I want to mention/make note of the messages not remaining in the ESv10 UI issue and that with the Consumer binaries you can still properly consume the messages and that they haven't magically disappeared.