Updated IBM COS usecase (#110)
* Updated IBM COS usecase

Clarified the sections for Kafka Connect Cluster and the COS SInk connector that those sections are on the previous ES version and that there's an ESv10 section. Also emphasized the point that the COS Sink connector needs to be built with Java 8.

* Update index.mdx

Reverted some steps and opted to have all the outlined steps cover the older Event Streams version and to just have the ESv10 section outline all the new screenshots and yamls instead.